Module amazon_sp_feeds::apis::feeds_api

Enums

CancelFeedError

struct for typed errors of method cancel_feed

CreateFeedDocumentError

struct for typed errors of method create_feed_document

CreateFeedError

struct for typed errors of method create_feed

GetFeedDocumentError

struct for typed errors of method get_feed_document

GetFeedError

struct for typed errors of method get_feed

GetFeedsError

struct for typed errors of method get_feeds

Functions

cancel_feed

Cancels the feed that you specify. Only feeds with processingStatus=IN_QUEUE can be cancelled. Cancelled feeds are returned in subsequent calls to the getFeed and getFeeds operations. Usage Plan: | Rate (requests per second) | Burst | | –– | –– | | 0.0222 | 10 | For more information, see "Usage Plans and Rate Limits" in the Selling Partner API documentation.

create_feed

Creates a feed. Encrypt and upload the contents of the feed document before calling this operation. Usage Plan: | Rate (requests per second) | Burst | | –– | –– | | 0.0083 | 15 | For more information, see "Usage Plans and Rate Limits" in the Selling Partner API documentation.

create_feed_document

Creates a feed document for the feed type that you specify. This operation returns encryption details for encrypting the contents of the document, as well as a presigned URL for uploading the encrypted feed document contents. It also returns a feedDocumentId value that you can pass in with a subsequent call to the createFeed operation. Usage Plan: | Rate (requests per second) | Burst | | –– | –– | | 0.0083 | 15 | For more information, see "Usage Plans and Rate Limits" in the Selling Partner API documentation.

get_feed

Returns feed details (including the resultDocumentId, if available) for the feed that you specify. Usage Plan: | Rate (requests per second) | Burst | | –– | –– | | 2.0 | 15 | For more information, see "Usage Plans and Rate Limits" in the Selling Partner API documentation.

get_feed_document

Returns the information required for retrieving a feed document’s contents. This includes a presigned URL for the feed document as well as the information required to decrypt the document’s contents. Usage Plan: | Rate (requests per second) | Burst | | –– | –– | | 0.0222 | 10 | For more information, see "Usage Plans and Rate Limits" in the Selling Partner API documentation.

get_feeds

Returns feed details for the feeds that match the filters that you specify. Usage Plan: | Rate (requests per second) | Burst | | –– | –– | | 0.0222 | 10 | For more information, see "Usage Plans and Rate Limits" in the Selling Partner API documentation.